Replace pa-info test (just a shell script) with select commands
A lot of the pa-info things fail anyway because of missing tools (e.g.
`ps` or `which`), dependency tools (jack, alsa), and files at global
locations (/etc) that don't get installed by the conda package. Instead,
pull out a couple of the commands that pa-info was running that are
specific to the pulseaudio package and run only those.